:lipstick: create style files for EmotionCard
